Assignment 03-1
Massive injection of silver iodide into cumulus clouds may lead to increased rainfall. To test this hypothesis, a flight experiment is conducted. On each of 52 days that were deemed suitable for cloud seeding, a random mechanism was used to decide whether to seed the target cloud on that day or leave it unseeded as a control.

An airplane flew through the cloud in both cases, since the experimenters and the pilot were themselves unaware of whether on any particular day the seeding mechanism in the plane was loaded or not (that is, they were blind to the treatment). Precipitation was measured as the total rain volume falling from the cloud base following the airplane seeding run, as measured by radar.