Assignment 03-4
4. Atmospheric Lapse Rate
The lapse rate is the rate at which temperature drops as you increase elevation. Some hardy students were interested in checking empirically if the lapse rate of 9.8 degrees C km-1 was accurate for their hiking. To investigate, they grabbed their thermometers and their Huawei wrist altimeters and found the following data on their hike.Draw a scatter plot with regression line, and investigate if the lapse rate is 9.8 degrees C km-1.