Assignment 03-5
5. The Big Bang Theory
Edwin Hubble used the power of the Mount Wilson Observatory telescopes to measure features of the nebulae outside the Milky Way. He was surprised to find a relationship between a nebula’s distance from the Earth and the velocity with which is was going away from the Earth. Hubble’s initial data on 24 nebulae are listed below.